Implementar endpoint ROLE_ADMIN para calcular y mostrar reporte de ganancias generales del centro deportivo. Agregar dos nuevas excepciones y corrección del endpoint para empezar reserva. Nuevo TurnoState.COMPLETADO.

+++ src/main/java/com/project/deporturnos/entity/dto/CanchaResponseDTO.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @AllArgsConstructor
@@ -18,7 +20,7 @@ public class CanchaResponseDTO {
 
     private String tipo;
 
-    private double precioHora;
+    private BigDecimal precioHora;
 
     private boolean disponibilidad;
